+++ concordium-consensus/src/Concordium/External.hs
@@ -311,6 +311,8 @@ startConsensus ::
     Word64 ->
     -- |Number of seconds between transaction table purging runs
     Word64 ->
+    -- |Accounts table cache size
+    Word32 ->
     -- |Serialized genesis data (c string + len)
     CString ->
     Int64 ->
@@ -344,6 +346,7 @@ startConsensus
     insertionsBeforePurge
     transactionsKeepAlive
     transactionsPurgingDelay
+    accountsCacheSize
     gdataC
     gdataLenC
     bidC
@@ -415,7 +418,8 @@ startConsensus
                   rpMaxBakingDelay = defaultMaxBakingDelay,
                   rpInsertionsBeforeTransactionPurge = fromIntegral insertionsBeforePurge,
                   rpTransactionsKeepAliveTime = TransactionTime transactionsKeepAlive,
-                  rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay
+                  rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay,
+                  rpAccountsCacheSize = fromIntegral accountsCacheSize
                 }
 
 -- |Start up an instance of Skov without starting the baker thread.
@@ -432,6 +436,8 @@ startConsensusPassive ::
     Word64 ->
     -- |Number of seconds between transaction table purging runs
     Word64 ->
+    -- |Accounts table cache size
+    Word32 ->
     -- |Serialized genesis data (c string + len)
     CString ->
     Int64 ->
@@ -460,6 +466,7 @@ startConsensusPassive
     insertionsBeforePurge
     transactionsKeepAlive
     transactionsPurgingDelay
+    accountsCacheSize
     gdataC
     gdataLenC
     cucbk
@@ -513,7 +520,8 @@ startConsensusPassive
                   rpMaxBakingDelay = defaultMaxBakingDelay,
                   rpInsertionsBeforeTransactionPurge = fromIntegral insertionsBeforePurge,
                   rpTransactionsKeepAliveTime = TransactionTime transactionsKeepAlive,
-                  rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay
+                  rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay,
+                  rpAccountsCacheSize = fromIntegral accountsCacheSize
                 }
 
 -- |Shut down consensus, stopping any baker thread if necessary.
@@ -1220,6 +1228,8 @@ foreign export ccall
         Word64 ->
         -- |Number of seconds between transaction table purging runs
         Word64 ->
+        -- |Accounts table cache size
+        Word32 ->
         -- |Serialized genesis data (c string + len)
         CString ->
         Int64 ->
@@ -1259,6 +1269,8 @@ foreign export ccall
         Word64 ->
         -- |Number of seconds between transaction table purging runs
         Word64 ->
+        -- |Accounts table cache size
+        Word32 ->
         -- |Serialized genesis data (c string + len)
         CString ->
         Int64 ->