+++ concordium-node/src/configuration.rs
@@ -263,6 +263,13 @@ pub struct BakerConfig {
     )]
     pub genesis_data_file: PathBuf,
     #[structopt(
+        long = "accounts-cache-size",
+        help = "The maximum number of accounts that can be stored in accounts cache",
+        default_value = "10000",
+        env = "CONCORDIUM_NODE_CONSENSUS_ACCOUNTS_CACHE_SIZE"
+    )]
+    pub account_cache_size: u32,
+    #[structopt(
         long = "baker-credentials-file",
         help = "Path to the baker credentials file. If the path is relative it is interpreted \
                 relative to the node process' working directory.",