+++ concordium-node/src/consensus_ffi/ffi.rs
@@ -259,8 +259,8 @@ extern "C" {
         max_block_size: u64,
         block_construction_timeout: u64,
         insertions_before_purging: u64,
-        transaction_keep_alive: u64,
         transactions_purging_delay: u64,
+        accounts_cache_size: u32,
         genesis_data: *const u8,
         genesis_data_len: i64,
         private_data: *const u8,
@@ -282,6 +282,7 @@ extern "C" {
         insertions_before_purging: u64,
         transaction_keep_alive: u64,
         transactions_purging_delay: u64,
+        accounts_cache_size: u32,
         genesis_data: *const u8,
         genesis_data_len: i64,
         catchup_status_callback: CatchUpStatusCallback,
@@ -452,14 +453,8 @@ extern "C" {
     pub fn stopImportingBlocks(consensus: *mut consensus_runner);
 }
 
-// TODO : Simplify arguments to function, or group with struct
-#[allow(clippy::too_many_arguments)]
 pub fn get_consensus_ptr(
-    max_block_size: u64,
-    block_construction_timeout: u64,
-    insertions_before_purging: u64,
-    transaction_keep_alive: u64,
-    transactions_purging_delay: u64,
+    runtime_parameters: &ConsensusRuntimeParameters,
     genesis_data: Vec<u8>,
     private_data: Option<Vec<u8>>,
     maximum_log_level: ConsensusLogLevel,
@@ -476,11 +471,11 @@ pub fn get_consensus_ptr(
             let appdata_buf = appdata_dir.to_str().unwrap();
             unsafe {
                 startConsensus(
-                    max_block_size,
-                    block_construction_timeout,
-                    insertions_before_purging,
-                    transaction_keep_alive,
-                    transactions_purging_delay,
+                    runtime_parameters.max_block_size,
+                    runtime_parameters.block_construction_timeout,
+                    runtime_parameters.insertions_before_purging,
+                    runtime_parameters.transactions_purging_delay,
+                    runtime_parameters.accounts_cache_size,
                     genesis_data.as_ptr(),
                     genesis_data_len as i64,
                     private_data_bytes.as_ptr(),
@@ -503,11 +498,12 @@ pub fn get_consensus_ptr(
             unsafe {
                 {
                     startConsensusPassive(
-                        max_block_size,
-                        block_construction_timeout,
-                        insertions_before_purging,
-                        transaction_keep_alive,
-                        transactions_purging_delay,
+                        runtime_parameters.max_block_size,
+                        runtime_parameters.block_construction_timeout,
+                        runtime_parameters.insertions_before_purging,
+                        runtime_parameters.transaction_keep_alive,
+                        runtime_parameters.transactions_purging_delay,
+                        runtime_parameters.accounts_cache_size,
                         genesis_data.as_ptr(),
                         genesis_data_len as i64,
                         catchup_status_callback,